refactor: organize types into 'types' folder
