+++ src/data/types/index.ts
@@ -1,64 +1,5 @@
 import { SwapPegTransaction } from '@/domain/entities/report.entity';
 
-export interface User {
-  id: string;
-  username: string;
-  email?: string;
-  name?: string;
-  providerId?: string;
-  isActive: boolean;
-  level: number;
-  refreshToken?: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface UserFilter {
-  username?: string;
-  status?: 'all' | 'active' | 'inactive';
-  level?: number;
-}
-
-export interface Coupon {
-  id: string;
-  code: string;
-  discountType: 'percentage' | 'fixed';
-  discountValue: number;
-  minPurchaseValue: number;
-  maxDiscountValue: number;
-  usageLimit: number;
-  usedCount: number;
-  isActive: boolean;
-  validFrom: string;
-  validUntil: string | null | undefined;
-}
-
-export interface CouponFilter {
-  code: string;
-  status: 'all' | 'active' | 'inactive';
-}
-
-export type CreateCouponDTO = Omit<Coupon, 'id' | 'usedCount'> & {
-  validUntil?: string | null;
-};
-
-export type UpdateCouponDTO = Partial<CreateCouponDTO>;
-
-export interface Sale {
-  id: string;
-  userId: string;
-  total: number;
-  status: 'pending' | 'completed' | 'cancelled';
-  paymentMethod: string;
-  createdAt: string;
-  items: {
-    id: string;
-    name: string;
-    quantity: number;
-    price: number;
-  }[];
-}
-
 export interface DashboardStats {
   totalSales: number;
   totalRevenue: number;
