+++ src/data/types/userTypes.ts
@@ -1,0 +1,18 @@
+export interface User {
+  id: string;
+  username: string;
+  email?: string;
+  name?: string;
+  providerId?: string;
+  isActive: boolean;
+  level: number;
+  refreshToken?: string;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface UserFilter {
+  username?: string;
+  status?: 'all' | 'active' | 'inactive';
+  level?: number;
+}
